Fixed the project structure, fixed tests, example use in py, will be revized into th .ipynb later.
