+++ .gitignore
@@ -135,3 +135,4 @@ cache/
 #semantic_models
 /.model
 .model
+/semantic_models
