--- semantic_vector/data/output/graph.csv
@@ -1,111 +1,0 @@
-subject;object;semantic_closeness
-тротуар;дорожка;0.9089667797088623
-тротуар;пешеходный;0.8945724368095398
-тротуар;парковка;0.881375789642334
-тротуар;проход;0.8813527822494507
-тротуар;дорога;0.8789024949073792
-тротуар;расчистить;0.8763077259063721
-тротуар;сквер;0.8736048340797424
-тротуар;газон;0.8716099858283997
-тротуар;забор;0.8710044622421265
-тротуар;вдоль;0.868878185749054
-дорожка;пешеходный;0.9493331909179688
-дорожка;расчистить;0.933650553226471
-дорожка;тропинка;0.9235348701477051
-дорожка;проход;0.909407377243042
-дорожка;тротуар;0.9089667797088623
-дорожка;пешеход;0.9037609100341797
-дорожка;сквер;0.8965794444084167
-дорожка;врезаться;0.8926591873168945
-дорожка;забор;0.8892408013343811
-дорожка;куст;0.8871298432350159
-расчистить;дорожка;0.9336504936218262
-расчистить;парковаться;0.9326648116111755
-расчистить;сквер;0.9322994947433472
-расчистить;завалить;0.9300934672355652
-расчистить;тропинка;0.9196770787239075
-расчистить;поребрик;0.9182372689247131
-расчистить;узкий;0.9171484708786011
-расчистить;врезаться;0.9153966903686523
-расчистить;легковой;0.9153388738632202
-расчистить;проход;0.9111073017120361
-сквер;тропинка;0.9414651393890381
-сквер;расчистить;0.9322995543479919
-сквер;сплошной;0.9265956878662109
-сквер;путь;0.9219260811805725
-сквер;поребрик;0.9216931462287903
-сквер;месиво;0.9190894961357117
-сквер;огородить;0.9185354709625244
-сквер;муринский;0.9167721271514893
-сквер;пята;0.9131755232810974
-сквер;каршеринг;0.9093141555786133
-путь;влететь;0.951842188835144
-путь;каршеринг;0.944101870059967
-путь;зел;0.9430592656135559
-путь;превращаться;0.9427375197410583
-путь;врезаться;0.9404396414756775
-путь;самосвал;0.940253496170044
-путь;поребрик;0.9399515986442566
-путь;налево;0.9394518136978149
-путь;поворот;0.9388467669487
-путь;светофор;0.9384920597076416
-врезаться;каршеринг;0.988109290599823
-врезаться;газель;0.981462836265564
-врезаться;легковушка;0.9799216389656067
-врезаться;бмв;0.9753094911575317
-врезаться;киа;0.9743484258651733
-врезаться;фольксваген;0.9737027883529663
-врезаться;шкода;0.9721469283103943
-врезаться;влететь;0.971462070941925
-врезаться;мерседес;0.9706683158874512
-врезаться;легковой;0.9704341292381287
-фольксваген;лада;0.992791473865509
-фольксваген;форд;0.9926639199256897
-фольксваген;рено;0.9886134266853333
-фольксваген;микроавтобус;0.9876585006713867
-фольксваген;шкода;0.9875303506851196
-фольксваген;бмв;0.986311137676239
-фольксваген;легковушка;0.9861711263656616
-фольксваген;ниссан;0.9825902581214905
-фольксваген;газель;0.9818426966667175
-фольксваген;перевернуться;0.9813291430473328
-легковушка;газель;0.9936932325363159
-легковушка;бмв;0.9924497604370117
-легковушка;форд;0.9916864633560181
-легковушка;каршеринг;0.989689290523529
-легковушка;мерседес;0.9892382621765137
-легковушка;шкода;0.9876124858856201
-легковушка;киа;0.9874897599220276
-легковушка;мотоцикл;0.9866203665733337
-легковушка;фольксваген;0.986171305179596
-легковушка;солярис;0.9850751757621765
-мотоцикл;бмв;0.9956157207489014
-мотоцикл;шкода;0.9941561818122864
-мотоцикл;форд;0.9935102462768555
-мотоцикл;мерседес;0.9934114813804626
-мотоцикл;солярис;0.9932534694671631
-мотоцикл;киа;0.9927595257759094
-мотоцикл;встречка;0.9886717200279236
-мотоцикл;полый;0.9885918498039246
-мотоцикл;легковушка;0.9866202473640442
-мотоцикл;уступить;0.986102283000946
-легковушка;газель;0.9936932325363159
-легковушка;бмв;0.9924497604370117
-легковушка;форд;0.9916864633560181
-легковушка;каршеринг;0.989689290523529
-легковушка;мерседес;0.9892382621765137
-легковушка;шкода;0.9876124858856201
-легковушка;киа;0.9874897599220276
-легковушка;мотоцикл;0.9866203665733337
-легковушка;фольксваген;0.986171305179596
-легковушка;солярис;0.9850751757621765
-солярис;мерседес;0.996196985244751
-солярис;бмв;0.99462890625
-солярис;мотоцикл;0.9932534694671631
-солярис;лазурный;0.9910354614257812
-солярис;киа;0.9882804751396179
-солярис;встречка;0.9879289269447327
-солярис;шкода;0.9872850775718689
-солярис;форд;0.9870221614837646
-солярис;уступить;0.9863559007644653
-солярис;влететь;0.9863530993461609
