--- semantic_vector/src/SEMANTIC_README.md
@@ -1,24 +1,0 @@
-Модуль для построения семантического графа на основе текстов (.csv или .txt)
-
-**Требует Python < 3.10**
-
-Структура:
-
-* src - исполняемы модули проекта
-    * paths.py - содержит пути к папкам
-    * preprocessor.py - описан класс для очистки текста и приведения слов в начальную форму
-    * train_model.py - описан класс для создания, тренировки и сохранения модели
-    * visualization.py - создаёт граф на основе входного слова и параметров количества ближайших по смыслу слов и глубины
-    * use_example.py - пример использования классов с параметрками по умолчанию
-
-* models - хранилище моделей
-    * custom - польовательские модели
-
-* data - хранилище данных
-    * train - тренировочные
-        * raw - непредобработанные данные
-    * output - выходные данные, в текущей момент служет хранилищем для csv из которого строится семантический граф
-
-* kg1.jpg - файл с изображением созданного графа
-
-В коде есть пояснительные комментарии.
